Add oh-my-zsh-compatible plugin file.
Add ugit.plugin.zsh. This is the standard name used by Oh-My-ZSH
(and oh-my-zsh-compatible ZSH frameworks) for plugins, and once
added to the user's configuration, will be loaded automatically
during login, and add the repo directory to the user's `$PATH` to
make `ugit` available.

This is also used by smarter ZSH frameworks like Antigen and Zgenom
to automatically keep the user's checkout of `ugit` up to date so
they get automatic updates.

Closes #10